# docs/swagger.yaml
definitions:
...
  routes.StatusRes:
    properties:
      status:
        example: status of the request
        type: string
    type: object
...
  routes.getMetadataRes:
    properties:
      expirationDate:
        type: string
      metadata:
        example: your account metadata
        type: string
    required:
    - expirationDate
    - metadata
    type: object
  routes.getMetadataV2Res:
    properties:
      expirationDate:
        type: string
      metadataV2:
        example: your account metadataV2
        type: string
    required:
    - expirationDate
    - metadataV2
    type: object
...
  routes.metadataKeyReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.metadataMultipleV2KeyReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.metadataV2KeyReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.metadataV2PublicKeyReq:
    properties:
      requestBody:
        example: look at description for example
        type: string
    required:
    - requestBody
    type: object
...
  routes.updateMetadataV2Req:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.updateMetadataV2Res:
    properties:
      expirationDate:
        type: string
      metadataV2:
        example: your (updated) account metadataV2
        type: string
      metadataV2Key:
        example: public key for the metadataV2 encoded to base64url
        type: string
    required:
    - expirationDate
    - metadataV2
    - metadataV2Key
    type: object
...
info:
  contact:
    name: Opacity Staff
    url: https://telegram.me/opacitystorage
  description: Opacity backend for file storage.
  license:
    name: OPACITY LIMITED CODE REVIEW LICENSE LICENSE.md
  termsOfService: https://opacity.io/terms-of-service
  title: Storage Node
  version: "1.0"
paths:
...
  /api/v1/metadata/delete:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataKey": "a 64-char hex string created deterministically, will be a key for the metadata of one of your folders",
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoints that only need metadataKey and timestamp
        in: body
        name: metadataKeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: delete a metadata
  /api/v1/metadata/get:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataKey": "a 64-char hex string created deterministically, will be a key for the metadata of one of your folders",
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoints that only need metadataKey and timestamp
        in: body
        name: metadataKeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.getMetadataRes'
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: no value found for that key, or account not found
          schema:
            type: string
      summary: Retrieve account metadata
...
  /api/v2/metadata/add:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataV2Key": "public key for the metadataV2 encoded to base64",
        "metadataV2Vertex": "the vertex to add to your account metadataV2 encoded to base64",
        "metadataV2Edges": "the edges to add to your account metadataV2 encoded to base64",
        "metadataV2Sig": "a signature encoded to base64 confirming the metadata change, the publickey will be a key for the metadataV2",
        "timestamp": 1557346389
        }
      parameters:
      - description: update metadataV2 object
        in: body
        name: updateMetadataV2Req
        required: true
        schema:
          $ref: '#/definitions/routes.updateMetadataV2Req'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.updateMetadataV2Res'
        "400":
          description: 'bad request, can''t verify signature: (with the error)'
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice response
          schema:
            type: string
        "404":
          description: no value found for that key, or account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: Update metadataV2
  /api/v2/metadata/delete:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataV2Key": "public key for the metadataV2 encoded to base64",
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoints that only need metadataV2Key and timestamp
        in: body
        name: metadataV2KeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataV2KeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: bad request, incorrect key length
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: delete a metadataV2
  /api/v2/metadata/delete-multiple:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataV2Keys": ["public key for the metadataV2 encoded to base64", "another public key for the metadataV2 encoded to base64", "..."],
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoint that only needs an array of metadataV2Keys
          and timestamp
        in: body
        name: metadataMultipleV2KeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataMultipleV2KeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: bad request, incorrect key length
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: account not found
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: delete multiple metadataV2, if a key is not found, it won't be treated
        as an error
  /api/v2/metadata/get:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataV2Key": "public key for the metadataV2 encoded to base64",
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoints that only need metadataV2Key and timestamp
        in: body
        name: metadataV2KeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataV2KeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.getMetadataV2Res'
        "400":
          description: bad request, incorrect key length
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: no value found for that key, or account not found
          schema:
            type: string
      summary: Retrieve account metadataV2
  /api/v2/metadata/get-public:
    post:
      consumes:
      - application/json
      description: |-
        requestBody should be a stringified version of (values are just examples):
        {
        "metadataV2Key": "public key for the metadataV2 encoded to base64",
        "timestamp": 1557346389
        }
      parameters:
      - description: object for endpoints that only need metadataV2Key and timestamp
        in: body
        name: metadataV2PublicKeyReq
        required: true
        schema:
          $ref: '#/definitions/routes.metadataV2PublicKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.getMetadataV2Res'
        "400":
          description: bad request, incorrect key length
          schema:
            type: string
        "403":
          description: subscription expired, or the invoice resonse
          schema:
            type: string
        "404":
          description: no value found for that key, or account not found
          schema:
            type: string
      summary: Retrieve account metadataV2
...
swagger: "2.0"